Refactored tag-related JS for reuseability, and added map tags support.
