+++ source/PopNGo/PopNGo/PopNGo/Views/Map/Index.cshtml
@@ -11,6 +11,7 @@
         <p id="date"></p>
         <p id="description"></p>
         <img id="image" src="" alt="Event Image">
+        <div id="tags" class="px-3 tags"></div>
     </div>
     <div id="demo-map-id" style="width: 100vw;">
         <!-- Map goes here -->
