+++ source/PopNGo/PopNGo/PopNGo/wwwroot/js/eventsAPI.js
@@ -31,6 +31,20 @@ export async function fetchEventData(query) {
     }
 }
 
+export async function searchForEvents(query, callback) {
+    const searchQuery = query ?? document.getElementById('search-event-input').value;
+    document.getElementById('no-events-section')?.classList.toggle('hidden', true); // Hide the no events section
+    document.getElementById('searching-events-section')?.classList.toggle('hidden', false); // Show the searching events section
+
+    if (searchQuery) {
+        fetchEventData(searchQuery).then(data => {
+            callback(data); // Assuming the data structure includes an array in data.data
+        }).catch(e => {
+            console.error('Fetching events failed:', e);
+        });
+    }
+}
+
 export async function fetchTagId(tag) {
     try {
         const response = await fetch(`/api/tags/name=${tag}`);
@@ -45,16 +59,28 @@ export async function fetchTagId(tag) {
     }
 }
 
-export async function createTags(tagList) {
+export async function createTags(events) {
+    if(events.length === 0) return;
+
+    // Create a set of unique tags across the events
+    let tagList = new Set();
+    events?.forEach(event => {
+        event.eventTags?.forEach(tag => {
+            tag = capitalize(tag).replace(/-|_/g, ' ');
+            tagList.add(tag);
+        });
+    });
+
     if(tagList.length === 0) return;
 
+    // Create the tags
     try {
         const response = await fetch(`/api/tags/create`, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
             },
-            body: JSON.stringify(tagList)
+            body: JSON.stringify(Array.from(tagList))
         });
         if (!response.ok) {
             throw new Error('Network response was not ok');
@@ -65,16 +91,38 @@ export async function createTags(tagList) {
     }
 }
 
-export async function searchForEvents(query, callback) {
-    const searchQuery = query ?? document.getElementById('search-event-input').value;
-    document.getElementById('no-events-section')?.classList.toggle('hidden', true); // Hide the no events section
-    document.getElementById('searching-events-section')?.classList.toggle('hidden', false); // Show the searching events section
+export async function formatTags(event, tagsParent) {
+    processArray(event.eventTags, async (tag) => {
+        tag = capitalize(tag).replace(/-|_/g, ' ');
 
-    if (searchQuery) {
-        fetchEventData(searchQuery).then(data => {
-            callback(data); // Assuming the data structure includes an array in data.data
-        }).catch(e => {
-            console.error('Fetching events failed:', e);
-        });
-    }
+        const tagEl = document.createElement('span');
+        tagEl.classList.add('tag');
+
+        let tagIndex = await fetchTagId(tag) || null;
+        if(tagIndex !== null)
+            tagEl.classList.add(`tag-${tagIndex}`);
+
+        tagEl.textContent = tag;
+        tagsParent.appendChild(tagEl);
+    }).then(() => {
+        // Grab the children we just made
+        let children = Array.prototype.slice.call(tagsParent.children);
+
+        // Sort the children elements
+        children.sort((a, b) => a.textContent.localeCompare(b.textContent));
+
+        // Append each child back to tags
+        children.forEach(child => tagsParent.appendChild(child));
+    });
+}
+
+export async function processArray(array, asyncFunction) {
+    // map array to promises
+    const promises = array.map(asyncFunction);
+    // wait until all promises resolve
+    await Promise.all(promises);
+}
+
+export function capitalize(str) {
+    return str.charAt(0).toUpperCase() + str.slice(1);
 }
